Rename composable to private
